+++ src/project/manager.js
@@ -8,8 +8,13 @@
 
 import * as pioNodeHelpers from 'pioarduino-node-helpers';
 import * as projectHelpers from './helpers';
-
 import { disposeSubscriptions, notifyError } from '../utils';
+import {
+  ensureClangdArgs,
+  fixupCompileCommands,
+  getActiveBackend,
+  notifyRescanBackend,
+} from '../intellisense';
 import { ProjectConfigLanguageProvider } from './config';
 import ProjectTaskManager from './tasks';
 import ProjectTestManager from './tests';
@@ -30,8 +35,10 @@ export default class ProjectManager {
     this._configProvider = new ProjectConfigLanguageProvider();
     this._configChangedTimeout = undefined;
 
+    const activeBackend = getActiveBackend();
     this._pool = new pioNodeHelpers.project.ProjectPool({
-      ide: 'vscode',
+      ide: activeBackend.indexerIde,
+      intelliSenseBackend: activeBackend,
       api: {
         logOutputChannel: this._logOutputChannel,
         createFileSystemWatcher: vscode.workspace.createFileSystemWatcher,
@@ -83,6 +90,11 @@ export default class ProjectManager {
           );
         },
         onDidNotifyError: notifyError.bind(this),
+        onDidRebuildIndex: async (projectDir) => {
+          await fixupCompileCommands(projectDir);
+          await ensureClangdArgs(projectDir);
+          await notifyRescanBackend();
+        },
       },
       settings: {
         autoPreloadEnvTasks: extension.getConfiguration('autoPreloadEnvTasks'),
@@ -227,6 +239,7 @@ export default class ProjectManager {
     ) {
       disposeSubscriptions(this.internalSubscriptions);
       await this._pool.switch(projectDir);
+      await ensureClangdArgs(projectDir);
       this._taskManager = new ProjectTaskManager(projectDir, observer);
       this.internalSubscriptions.push(
         this._taskManager,